refactor: move detection service and model to root package and update imports and configuration
