+++ detection/src/main/java/pt/estga/detection/config/DetectionAutoConfiguration.java
@@ -1,7 +1,6 @@
 package pt.estga.detection.config;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.persistence.autoconfigure.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -9,8 +8,7 @@ import org.springframework.web.client.RestTemplate;
 
 @AutoConfiguration
 @ComponentScan("pt.estga.detection")
-@EntityScan("pt.estga.detection.entities")
-@EnableConfigurationProperties
+@EntityScan("pt.estga.detection")
 public class DetectionAutoConfiguration {
 
     @Bean
