+++ submission/src/main/java/pt/estga/submission/listeners/SubmissionEnrichmentListener.java
@@ -8,8 +8,8 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
-import pt.estga.detection.model.DetectionResult;
-import pt.estga.detection.service.DetectionService;
+import pt.estga.detection.DetectionResult;
+import pt.estga.detection.DetectionService;
 import pt.estga.file.services.MediaService;
 import pt.estga.submission.events.SubmissionSubmittedEvent;
 import pt.estga.submission.repositories.MarkOccurrenceSubmissionRepository;
